split read function into read collections and docs
