+++ src/lib/firebase/WriteToCloudFirestore.ts
@@ -1,5 +1,5 @@
 import { collection, addDoc, setDoc, doc } from 'firebase/firestore'
-import { useFirebaseFirestore } from './useFirebase'
+import { useFirebaseFirestore } from './hooks/useFirebase'
 import Table from '../../types/Table'
 import User from '../../types/User'
 
@@ -7,9 +7,12 @@ import User from '../../types/User'
  * writeToCloudFirestore(col: string, doc: string, data: any)
  * Handles writing data to cloud firestore
  * 
- * @param {string} col 
- * @param {*} data 
- * @param {string} doc 
+ * @param {string} col Collection name to write to
+ * @param {*} data Data to write
+ * @param {string} doc (optional) Document id to write to. If ommitted, assigns a random unique document id that is
+ * returned to user
+ * 
+ * @returns document id of written document
  */
 const WriteToCloudFirestore = async (collectionName: string, data: Table | User , docName? : string | undefined) => {
     const firestore = useFirebaseFirestore()
@@ -22,14 +25,16 @@ const WriteToCloudFirestore = async (collectionName: string, data: Table | User 
         if (docName) {
             await setDoc(doc(firestore, collectionName, docName!), data);
             console.log("Document written with ID: ", docName);
+            return docName
         }
         else {
             const docRef = await addDoc(collection(firestore, collectionName), data);
             console.log("Document written with ID: ", docRef.id);
+            return docRef.id
         }
       } catch (e) {
         console.error("Error adding document: ", e);
       }
 }
 
-export default WriteToCloudFirestore
+export { WriteToCloudFirestore }
